Fixed all of the tests.
The issue was I was using for in instead of for of. For in iterates through array, which means if unspecified you get an array back instead of the value. For of iterates and returns the actual value. I'm not sure why for of acts differently for arrays, but yet another TIL.